Refactored code and addressed comments.
